修改评论删除时 弹窗操作

+++ src/components/Singleblogs.vue
@@ -103,15 +103,47 @@
         });
       },
       // 评论删除
-      del(id){
-        this.$http.delete('article_comment/article_comment_id=' + id).then(response => {
-          alert("删除成功！")
-          this.reload()   // 调用页面刷新函数
-        }, response => {
-          console.log(response);
-          alert("出问题啦！")
+      // del(id){
+      //   this.$confirm('此操作将永久删除该评论, 是否继续?', '提示', {
+      //     confirmButtonText: '确定',
+      //     cancelButtonText: '取消',
+      //     type: 'warning'
+      //   }).then(() => {
+      //     this.$http.delete('article_comment/article_comment_id=' + id).then(response => {
+      //       alert("删除成功！")
+      //       this.reload()   // 调用页面刷新函数
+      //     });
+      //   }).catch(() => {
+      //     console.log(response);
+      //     alert("出问题啦！")
+      //   });
+      // },
+      del(id) {
+        this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          type: 'warning'
+        }).then(() => {
+          this.$http.delete('article_comment/article_comment_id=' + id).then(response => {
+            this.$message({
+              type: 'success',
+              message: '删除成功啦*  '
+            });
+            this.reload()   // 调用页面刷新函数
+          }, response => {
+            console.log(response);
+            this.$message({
+              type: 'info',
+              message: '出问题啦！ '
+            });
+          })
+        }).catch(() => {
+          this.$message({
+            type: 'info',
+            message: '已取消删除！ '
+          });
         })
-      },
+      }
     },
     mounted(){
       this.getArtComment()            //
